+++ src/properties/trees.c
@@ -503,28 +503,61 @@ static igraph_error_t igraph_i_is_forest(
  */
 igraph_error_t igraph_is_forest(const igraph_t *graph, igraph_bool_t *res,
                                 igraph_vector_int_t *roots, igraph_neimode_t mode) {
-    /* Caching is enabled only if the graph is undirected or mode == IGRAPH_ALL.
-     * Also, we can't return early if we need to calculate the roots */
-    igraph_bool_t use_cache = (
-        !igraph_is_directed(graph) || mode == IGRAPH_ALL
-    );
+
+    const igraph_bool_t treat_as_undirected = !igraph_is_directed(graph) || mode == IGRAPH_ALL;
 
     if (!roots && !res) {
         return IGRAPH_SUCCESS;
     }
 
-    if (use_cache && !roots && res) {
-        IGRAPH_RETURN_IF_CACHED_BOOL(graph, IGRAPH_PROP_IS_FOREST, res);
+    /* Note on cache use:
+     *
+     * The IGRAPH_PROP_IS_FOREST cached property is equivalent to this function's
+     * result ONLY in the undirected case. Keep in mind that a graph that is not
+     * a directed forest may still be an undirected forest, i.e. may still be free
+     * of undirected cycles. Example: 1->2<-3->4.
+     */
+
+    if (igraph_i_property_cache_has(graph, IGRAPH_PROP_IS_FOREST)) {
+        const igraph_bool_t no_undirected_cycles = igraph_i_property_cache_get_bool(graph, IGRAPH_PROP_IS_FOREST);
+        if (treat_as_undirected && res && ! roots) {
+            /* If the graph is treated as undirected and no roots are requested,
+             * we can directly use the cached IGRAPH_PROP_IS_FOREST value. */
+            *res = no_undirected_cycles;
+            return IGRAPH_SUCCESS;
+        } else {
+            /* Otherwise we can use negative cached values (i.e. "false"):
+             *  - A graph with undirected cycles cannot be a directed forest.
+             *  - If the graph is not a forest, we don't need to look for roots.
+             */
+            if (! no_undirected_cycles) {
+                if (res) { res = false; }
+                if (roots) { igraph_vector_int_clear(roots); }
+                return IGRAPH_SUCCESS;
+            }
+        }
     }
 
     IGRAPH_CHECK(igraph_i_is_forest(graph, res, roots, mode));
 
-    /* At this point we know whether the graph is a forest if we have at least
-     * one of 'res' or 'roots' */
-    if (res) {
-        igraph_i_property_cache_set_bool(graph, IGRAPH_PROP_IS_FOREST, *res);
-    } else if (roots) {
-        igraph_i_property_cache_set_bool(graph, IGRAPH_PROP_IS_FOREST, !igraph_vector_int_empty(roots));
+    /* At this point we know whether the graph is an (undirected or directed) forest
+     * as we have at least one of 'res' or 'roots'. The case when both are NULL was
+     * caught above. */
+    igraph_bool_t is_forest;
+    if (res != NULL) {
+        is_forest = *res;
+    } else /* roots != NULL */ {
+        is_forest = igraph_vcount(graph) == 0 || !igraph_vector_int_empty(roots);
+    }
+    if (is_forest) {
+        /* If the graph is a directed forest, then it has no undirected cycles.
+         * We can enter positive results in the cache unconditionally. */
+        igraph_i_property_cache_set_bool(graph, IGRAPH_PROP_IS_FOREST, true);
+    } else if (treat_as_undirected) {
+        /* However, if the graph is not a directed forest, it might still be
+         * an undirected forest. We can only enter negative results in the cache
+         * when edge directions were ignored, but NOT in the directed case. */
+        igraph_i_property_cache_set_bool(graph, IGRAPH_PROP_IS_FOREST, false);
     }
 
     return IGRAPH_SUCCESS;